pyproject.toml
[tool.poetry]
name = "chat-pdf"
version = "0.1.0"
description = ""
authors = ["anton-akulenko <anton.htl@gmail.com>"]
readme = "README.md"
packages = [{include = "chat_pdf"}]

[tool.poetry.dependencies]
python = "^3.10"
faiss-cpu = "^1.8.0"
langchain = "^0.1.12"
openai = "^1.14.1"
pypdfium2 = "^4.28.0"
tiktoken = "^0.6.0"
langchain-community = "^0.0.28"
pypdf2 = "^3.0.1"
langchain-openai = "^0.0.8"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
sentence-transformers = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
